refactor(rs): move json logic to backend
- Consolidate planar::Window with converters::Window into types::Window
- Move planar::Rect into types
- Replace to_window with From<&Value> for Window
