+++ rust/src/backend/i3/mod.rs
@@ -1,1 +1,2 @@
 pub mod client;
+pub mod compass;
