+++ rust/src/planar/mod.rs
@@ -1,9 +1,6 @@
 pub mod alignment;
 pub mod arrangement;
-pub mod rect;
 
 pub use alignment::Direction;
 pub use alignment::Relation;
 pub use arrangement::Arrangement;
-pub use rect::Rect;
-pub use arrangement::Window;
