+++ rust/src/types/mod.rs
@@ -1,0 +1,6 @@
+pub mod window;
+pub mod rect;
+
+pub use window::Window;
+pub use window::Windows;
+pub use rect::Rect;
